+++ DDD.Domain/ValueObjects/MeasureValue.cs
@@ -2,12 +2,12 @@
 {
     public sealed class MeasureValue : ValueObject<MeasureValue>
     {
-        public MeasureValue(float value)
+        public MeasureValue(double value)
         {
             Value = value;
         }
 
-        public float Value { get; }
+        public double Value { get; }
 
         public string DisplayValue => $"{Math.Round(Value, 2)}m/s";
